+++ Test/Database/PDO/TableReaderTest.php
@@ -2,19 +2,15 @@
 /**
  *	TestUnit of Database_PDO_TableReader.
  *	@package		Tests.database.pdo
- *	@extends		PHPUnit_Framework_TestCase
- *	@uses			Database_PDO_Connection
- *	@uses			Database_PDO_TableReader
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
  *	@since			02.07.2008
  *	@version		0.1
  */
-require_once 'PHPUnit/Framework/TestCase.php'; 
 require_once 'Test/initLoaders.php5';
 /**
  *	TestUnit of Database_PDO_TableReader.
  *	@package		Tests.database.pdo
- *	@extends		PHPUnit_Framework_TestCase
+ *	@extends		Test_Case
  *	@uses			Database_PDO_Connection
  *	@uses			Database_PDO_TableReader
  *	@author			Christian Würker <christian.wuerker@ceusmedia.de>
@@ -40,14 +36,6 @@ class Test_Database_PDO_TableReaderTest extends Test_Case
 		$this->queryLog	= $this->path."queries.log";
 
 		$this->dsn = "mysql:host=".$this->host.";dbname=".$this->database;
-
-		$options	= array();
-		
-		$this->connection	= new Database_PDO_Connection( $this->dsn, $this->username, $this->password, $options );
-		$this->connection->setAttribute( PDO::ATTR_CASE, PDO::CASE_NATURAL );
-		$this->connection->setAttribute( PDO::MYSQL_ATTR_USE_BUFFERED_QUERY, TRUE );
-		$this->connection->setErrorLogFile( $this->errorLog );
-		$this->connection->setStatementLogFile( $this->queryLog );
 		
 		$this->tableName	= "transactions";
 		$this->columns	= array(
@@ -62,7 +50,7 @@ class Test_Database_PDO_TableReaderTest extends Test_Case
 			'label'
 		);
 	}
-	
+
 	/**
 	 *	Setup for every Test.
 	 *	@access		public
@@ -70,6 +58,18 @@ class Test_Database_PDO_TableReaderTest extends Test_Case
 	 */
 	public function setUp()
 	{
+		if( !extension_loaded( 'pdo_mysql' ) )
+			$this->markTestSkipped( "PDO driver for MySQL not supported" );
+		if( !extension_loaded( 'mysql' ) )
+			$this->markTestSkipped( "Support for MySQL is missing" );
+
+		$options	= array();
+		$this->connection	= new Database_PDO_Connection( $this->dsn, $this->username, $this->password, $options );
+		$this->connection->setAttribute( PDO::ATTR_CASE, PDO::CASE_NATURAL );
+		$this->connection->setAttribute( PDO::MYSQL_ATTR_USE_BUFFERED_QUERY, TRUE );
+		$this->connection->setErrorLogFile( $this->errorLog );
+		$this->connection->setStatementLogFile( $this->queryLog );
+
 		$this->mysql	= mysql_connect( $this->host, $this->username, $this->password ) or die( mysql_error() );
 		mysql_select_db( $this->database );
 		$sql	= file_get_contents( $this->path."createTable.sql" );
@@ -90,7 +90,8 @@ class Test_Database_PDO_TableReaderTest extends Test_Case
 	{
 		@unlink( $this->errorLog );
 		@unlink( $this->queryLog );
-		mysql_query( "DROP TABLE transactions" );
+		if( extension_loaded( 'mysql' ) )
+			mysql_query( "DROP TABLE transactions" );
 	}
 
 	/**
